refactor: Simplify Orders class and integrate address processing
- Refactored OrdersImporter class into a more streamlined Orders class.
- Removed redundant methods related to product and order item processing, focusing on core order and address handling.
- Integrated address processing by utilizing the new Addresses class for improved clarity and maintainability.
- Enhanced error handling and logging during order processing, ensuring robust address ID retrieval.
- Cleaned up commented-out code to improve readability and focus on essential functionality.
